requirements.txt
faiss-cpu
flask
numpy
ollama
pypdf
python-dotenv
sentence-transformers
